Rename homeutils Go package to pathutils
